Product-Admin: Securing the views

+++ products/views.py
@@ -1,4 +1,5 @@
 from django.shortcuts import render, get_object_or_404, redirect, reverse
+from django.contrib.auth.decorators import login_required
 from django.contrib import messages
 from django.db.models import Q
 from django.db.models.functions import Lower
@@ -73,10 +74,15 @@ def product_detail(request, product_id):
     return render(request, "products/product_detail.html", context)    
 
 
+@login_required
 def add_product(request):
     """
     Add products to the store
     """
+    if not request.user.is_superuser:
+        messages.error(request, f"Sorry!, only store owner can do that.")
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -94,10 +100,15 @@ def add_product(request):
     return render(request, template, context=context)
 
 
+@login_required
 def edit_product(request, product_id):
     """
     Edit product in the store
     """
+    if not request.user.is_superuser:
+        messages.error(request, f"Sorry!, only store owner can do that.")
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -120,10 +131,15 @@ def edit_product(request, product_id):
     return render(request, template, context=context)
 
 
+@login_required
 def delete_product(request, product_id):
     """
     Delete product from the store
     """
+    if not request.user.is_superuser:
+        messages.error(request, f"Sorry!, only store owner can do that.")
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, f"Product Deleted")
